+++ yk_utils/objects/__init__.py
@@ -1,3 +1,2 @@
 """ Objects Module Init """
-from .object_operations import convert_to_string, datetime2str, \
-    get_url_domain, is_valid_url, get_url
+from .object_operations import convert_to_string, datetime2str